development/support-ai-assistant/backend/requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
openai
python-multipart
